+++ bind.go
@@ -106,11 +106,11 @@ func bind(server func(jid.JID, string) (jid.JID, error)) StreamFeature {
 			err = e.EncodeToken(start.End())
 			return req, err
 		},
-		Parse: func(ctx context.Context, r xml.TokenReader, start *xml.StartElement) (bool, interface{}, error) {
+		Parse: func(ctx context.Context, d *xml.Decoder, start *xml.StartElement) (bool, interface{}, error) {
 			parsed := struct {
 				XMLName xml.Name `xml:"urn:ietf:params:xml:ns:xmpp-bind bind"`
 			}{}
-			return true, nil, xml.NewTokenDecoder(r).DecodeElement(&parsed, start)
+			return true, nil, d.DecodeElement(&parsed, start)
 		},
 		Negotiate: func(ctx context.Context, session *Session, data interface{}) (mask SessionState, rw io.ReadWriter, err error) {
 			r := session.TokenReader()